scsi: sd: Ignore sync failures when not supported
Some external hard drives don't support the sync command even though the hard
drive has write cache enabled. In this case, ignore sync cache failures when the
error is for an illegal request for suspend. There's not much we can do for
these drives, so we shouldn't fail to suspend for this error case. The drive may
stay powered if that's setup for the port you have it plugged into.

BUG=chromium:282666
TEST=Plus Lacie hard drive into machine and make sure the machine can suspend.
	The Lacie hard drive may stay powered depending on the machine.

Change-Id: I88248f0715b70eae33eda171df3b803a5d7513b3
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/172863
Reviewed-by: Grant Grundler <grundler@chromium.org>

[benzh: 3.14 rebase. Merged error handling code for sd_sync_cache()]
[bfreed: 3.18 rebase. Fixed a compiler warning]
Signed-off-by: Ben Zhang <benzh@chromium.org>
Signed-off-by: Bryan Freed <bfreed@chromium.org>