UPSTREAM: cfg80211: allow RSSI compensation
Channels in 2.4GHz band overlap, this means that if we
send a probe request on channel 1 and then move to channel
2, we will hear the probe response on channel 2. In this
case, the RSSI will be lower than if we had heard it on
the channel on which it was sent (1 in this case).

The firmware / low level driver can parse the channel in
the DS IE or HT IE and compensate the RSSI so that it will
still have a valid value even if we heard the frame on an
adjacent channel. This can be done up to a certain offset.

Add this offset as a configuration for the low level driver.
A low level driver that can compensate the low RSSI in this
case should assign the maximal offset for which the RSSI
value is still valid.

Cherry-picked from commit 67af9811539be upstream.

Change-Id: Ib4c844e171fea3f822a203eb36f4532e0d3889c8
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/274893
Reviewed-by: Matt Chen <matt.chen@intel.com>
Commit-Queue: Matt Chen <matt.chen@intel.com>
Tested-by: Matt Chen <matt.chen@intel.com>
Reviewed-by: Paul Stewart <pstew@chromium.org>