CHROMIUM: soc/tegra: add suspend/resume for TSC
The clock source of the system counter (TSC) is not available when the
system suspends to deep sleep mode. It needs to be switched to the 32KHz
before suspending and restored after resuming.

The Tegra124/132 needs a SW WAR to avoid PMC wake status getting reset
when clearing TSC_MULT_EN in TSC resume code. To fix this, the ENABLE bit
also needs to be cleared at the same time.

BUG=chrome-os-partner:40741
TEST=LP0 suspend/resume test

Change-Id: I93d0f334cb7cb405c85015a72da2a6e99ca264d8
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/276950
Reviewed-by: Benson Leung <bleung@chromium.org>