FROMLIST: clk: tegra: Add functions for parsing CVB tables
Tegra CVB tables encode the relationship between operating voltage
and optimal frequency as a function of the so-called speedo value.
The speedo value is written to the on-chip fuses at the factory,
which allows the voltage-frequency operating points to be calculated
on an per-chip basis.

Add utility functions to parse the Tegra-specific tables and export the
voltage-frequency pairs to the generic OPP framework for other drivers
to use.

Signed-off-by: Tuomas Tynkkynen <ttynkkynen@nvidia.com>
Signed-off-by: Mikko Perttunen <mikko.perttunen@kapsi.fi>
Acked-by: Peter De Schrijver <pdeschrijver@nvidia.com>
Acked-by: Michael Turquette <mturquette@linaro.org>
(am from https://patchwork.kernel.org/patch/6398081/)

BUG=chrome-os-partner:40279
TEST=Build and boot

Change-Id: I6b48e2dfe04ec8d3557d54eb19a13abc78205fae
Signed-off-by: Penny Chiu <pchiu@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/271450
Reviewed-by: Benson Leung <bleung@chromium.org>