CHROMIUM: HACK: backlight/lp855x: Add tegra i2c6 hack
This patch adds a dummy i2c read in the lp855x probe to check
whether the bus is responding. If it is not, we defer probe
and try again.

This hack should be removed when the i2c bus initialization
depends on host1x power domain init.

Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/223454
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>

BUG=chrome-os-partner:39621
TEST=Tested on smaug

Change-Id: I4b583838ccea23d41d23a9a384dcf82865bc2419
Reviewed-on: https://chromium-review.googlesource.com/268302
Reviewed-by: David Riley <davidriley@chromium.org>
Commit-Queue: Sean Paul <seanpaul@google.com>
Tested-by: Sean Paul <seanpaul@google.com>