UPSTREAM: drivers: cpuidle: Add status property to ARM idle states
On some platforms the device tree bindings must provide the kernel
with a status flag for idle states, that defines whether the idle
state is operational or not in the current configuration.

This patch adds a status property to the ARM idle states compliant
with ePAPR v1.1 and updates the DT parsing code accordingly.

Acked-by: Kevin Hilman <khilman@linaro.org>
Signed-off-by: Lorenzo Pieralisi <lorenzo.pieralisi@arm.com>
Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>
(cherry picked from commit 97735da074fdd3fe39d975cad8760807df0d4388)

BUG=chrome-os-partner:39620
TEST=compile ok

Change-Id: I11c528517926ca600340e690a2ad3c08063ca8e8
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/269297
Reviewed-by: Benson Leung <bleung@chromium.org>