UPSTREAM: usb: gadget: f_midi: enable use of the index parameter
The soundcard index to use for the ALSA device creation is passed as a
parameter to f_midi_bind_config(), but is assigned to midi->index only
after the call to f_midi_register_card(midi). So no matter what is passed
to f_midi_bind_config(), the actual index for snd_card_new() is always 0.
This probably works ok if at the moment of f_midi's bind there are no
other snd_cards, but if there are, it is not possible to bind f_midi.

This patch moves the assignment to a place before the call to
f_midi_register_card(midi).

Signed-off-by: Andrzej Pietrasiewicz <andrzej.p@samsung.com>
Signed-off-by: Felipe Balbi <balbi@ti.com>
(cherry picked from commit 6e58ed578e2c65442b7a02deecc96e482c701654)
Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=b:23694613
TEST=Connect Smaug to a Linux workstation, switch to MIDI USB mode,
see the device with "aconnect -i" on the host, and see events with
"aseqdump -p NN".

Change-Id: If256f633c5c3984a8079d2e5c331fc190f7df677
Reviewed-on: https://chromium-review.googlesource.com/296764
Commit-Ready: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>