UPSTREAM: arm64/efi: move SetVirtualAddressMap() to UEFI stub
In order to support kexec, the kernel needs to be able to deal with the
state of the UEFI firmware after SetVirtualAddressMap() has been called.
To avoid having separate code paths for non-kexec and kexec, let's move
the call to SetVirtualAddressMap() to the stub: this will guarantee us
that it will only be called once (since the stub is not executed during
kexec), and ensures that the UEFI state is identical between kexec and
normal boot.

This implies that the layout of the virtual mapping needs to be created
by the stub as well. All regions are rounded up to a naturally aligned
multiple of 64 KB (for compatibility with 64k pages kernels) and recorded
in the UEFI memory map. The kernel proper reads those values and installs
the mappings in a dedicated set of page tables that are swapped in during
UEFI Runtime Services calls.

Acked-by: Leif Lindholm <leif.lindholm@linaro.org>
Acked-by: Matt Fleming <matt.fleming@intel.com>
Tested-by: Leif Lindholm <leif.lindholm@linaro.org>
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
(cherry picked from commit f3cdfd239da56a4cea75a2920dc326f0f45f67e3)
Signed-off-by: Tomasz Figa <tfiga@chromium.org>

Change-Id: I64d1f04caa7e72dcc4b0c344e5e99371ae9f1e74
Reviewed-on: https://chrome-internal-review.googlesource.com/209191
Reviewed-by: Benson Leung <bleung@chromium.org>
Tested-by: Tomasz Figa <tfiga@google.com>
Commit-Queue: Tomasz Figa <tfiga@google.com>
Reviewed-on: https://chromium-review.googlesource.com/267180
Tested-by: David Riley <davidriley@chromium.org>
Reviewed-by: David Riley <davidriley@chromium.org>
Commit-Queue: David Riley <davidriley@chromium.org>