UPSTREAM: ASoC: rt5677: Add TDM channel mapping function
It is for channel to slot mapping, and it is not only for 8 channels mapping,
but also in 2, 4 and 6 channels mapping. If we want to use the 2 channels in
the stereo2 adc path, we need to set the item "2/1/3/4" or "2/3/1/4".

It also adds for stereo channel swap. It can map the sterero channels "L/R"
to "R/L", "L/L" or R/R.

Signed-off-by: Oder Chiou <oder_chiou@realtek.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit e6f6ebc1f8f60d6d44f6be22c6386c238d6a9d97)

Signed-off-by: Koro Chen <koro.chen@mediatek.com>

BUG=chrome-os-partner:38746
TEST=build and boot oak-rev0

Change-Id: I0f96ff6393e65aed3adae7f1cdb638b889478050
Reviewed-on: https://chromium-review.googlesource.com/269033
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>
Trybot-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>