UPSTREAM: ASoC: dapm: Do not pretend to support controls for non mixer/mux widgets
Controls on a path only have an effect if the sink on the path is either a
mixer or mux widget. Currently we sort of silently ignore controls on other
paths, but since they don't do anything having them on other paths does not
make much sense and it is probably safe to assume that if we see such a path
it is a mistake in the driver that registered the path. This patch modifies
snd_soc_dapm_add_path() to report an error if a path with and control is
encountered where we didn't expect a control. This also allows to simplify
the code quite a bit.

The patch also moves the connecting of the path lists out of
dapm_connect_mux() and dapm_connect_mixer() into snd_soc_dapm_add_path().

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 5fe5b767dc6fb3df6fa6eaa8e05b727914f2bb4c)
BUG=chrome-os-partner:40036
TEST=Audio playback works on test device

Change-Id: I10ea6ff11daf9f1677c9bf07a683f62da1d3c60f
Reviewed-on: https://chromium-review.googlesource.com/270849
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Bernie Thompson <bhthompson@chromium.org>
Tested-by: Bernie Thompson <bhthompson@chromium.org>